fix(#6966): few more int64 casts

+++ src/maincurrencydialog.cpp
@@ -413,7 +413,7 @@ void mmMainCurrencyDialog::OnListItemSelected(wxDataViewEvent& event)
     if (is_selected)
     {
         wxDataViewItem item = event.GetItem();
-        m_currency_id = static_cast<int>(currencyListBox_->GetItemData(item));
+        m_currency_id = static_cast<int64>(currencyListBox_->GetItemData(item));
         Model_Currency::Data* currency = Model_Currency::instance().get(m_currency_id);
         if (currency)
         {
@@ -640,7 +640,7 @@ void mmMainCurrencyDialog::OnHistoryDelete(wxCommandEvent& WXUNUSED(event))
     {
         item = valueListBox_->GetNextItem(item, wxLIST_NEXT_ALL, wxLIST_STATE_SELECTED);
         if (item == -1) break;
-        Model_CurrencyHistory::instance().remove(static_cast<int>(valueListBox_->GetItemData(item)));
+        Model_CurrencyHistory::instance().remove(static_cast<int64>(valueListBox_->GetItemData(item)));
     }
     Model_CurrencyHistory::instance().ReleaseSavepoint();
